--- frontend/src/routes/demo/Demo.css
@@ -1,37 +1,0 @@
-*{
-    margin: 0;
-    padding: 0;
-    font-family: 'Times New Roman', Times, serif;
-}
-
-body{
-    background-color: #EFFBEE;
-}
-
-#body{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 80vh;
-}
-
-.conteudo{
-    padding: 20px;
-    background-color: white;
-    border-radius: 30px;
-    box-shadow: 0px 0px 20px black;
-}
-
-nav {
-    display: flex;
-    margin-top: 20px;
-}
-
-p{
-    margin: 0;
-}
-
-select {
-    margin-left: 1vw;
-    margin-right: 1vw;
-}
