+++ frontend/src/routes/pages/User/user.css
@@ -1,0 +1,7 @@
+body{
+    background-color: #EFFBEE;
+}
+
+#meusAlugueis{
+    background-color: #ffffff;
+}
